webhook/requirements.txt
Flask==3.1.0
gunicorn==23.0.0
kafka-python==2.1.3